Fixed Text Rendering, Integer Buffer Usage
The hack to use a floating point buffer to get text rendering working
produced plausible results on nvidia, but corrupted results on Intel.

This change corrects the texture upload for logging to use
GL_RED_INTEGER instead of GL_RED, and sets the format back to GL_R8I.

This also changes the ID buffer back to using an integer value, which
I changed for some reason earlier.  This works as expected on both
Intel and Nvidia.

+++ sdf_experiment.cpp
@@ -253,7 +253,7 @@ void AllocateRenderTargets(bool bErase=false)
 	glTextureParameteri(DepthBuffer, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE);
 
 	glCreateTextures(GL_TEXTURE_2D, 1, &ObjectIdBuffer);
-	glTextureStorage2D(ObjectIdBuffer, 1, GL_R32F, int(ScreenWidth), int(ScreenHeight));
+	glTextureStorage2D(ObjectIdBuffer, 1, GL_R32I, int(ScreenWidth), int(ScreenHeight));
 	glTextureParameteri(ObjectIdBuffer, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
 	glTextureParameteri(ObjectIdBuffer, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
 	glTextureParameteri(ObjectIdBuffer, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
